Move all resources exports to react-resource-router/resources
