+++ src/index.ts
@@ -73,31 +73,3 @@ export type {
 } from './controllers/router-store/types';
 
 export { DEFAULT_MATCH, DEFAULT_ROUTE } from './common/constants';
-
-// re-export resources entry-point to avoid breaking API changes for now
-export {
-  createResource,
-  ResourceDependencyError,
-  ResourceSubscriber,
-  useResource,
-  useResourceStoreContext,
-} from './resources';
-
-export type {
-  RouteResources,
-  ResourceStoreContext,
-  ResourceStoreData,
-  RouteResource,
-  RouteResourceError,
-  RouteResourceLoading,
-  RouteResourceResponse,
-  RouteResourceUpdater,
-  RouterDataContext,
-  UseResourceHookResponse,
-} from './common/types';
-
-export type {
-  CreateResourceArgSync,
-  CreateResourceArgAsync,
-  CreateResourceArgBase,
-} from './resources';
