+++ src/resources/controllers/resource-store/utils/dependent-resources/test.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { StoreActionApi } from 'react-sweet-state';
 
-import { ResourceType } from '../../../../../common/types';
+import { ResourceType } from '../../../../common/types';
 import {
   createResource,
   RouteResource,
   RouteResourceResponse,
-} from '../../../../../index';
+} from '../../../../index';
 import { State } from '../../types';
 
 import {
