+++ src/resources/controllers/resource-store/utils/route-checks/index.ts
@@ -1,7 +1,9 @@
-import { Route } from '../../../../../index';
+import { RouteWithResources } from '../../../../common/types';
 
-export const routeHasResources = (route: Route | null): boolean =>
+export const routeHasResources = (route: RouteWithResources | null): boolean =>
   !!(route && route.resources && route.resources.length > 0);
 
-export const routeHasChanged = (prev: Route, next: Route): boolean =>
-  prev.name !== next.name || prev.path !== next.path;
+export const routeHasChanged = (
+  prev: RouteWithResources,
+  next: RouteWithResources
+): boolean => prev.name !== next.name || prev.path !== next.path;
